address field problems on AuthType & expr
Two issues, they are sending an authorization type of EO instead of E0
(E zero, not capital O) and the month/year of the expiration is
reversed, should be mmyy it is coming across as 1411. The first one is
the biggie though, that's why it isn't even getting to first data.